requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
aiosqlite==0.20.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
playwright==1.48.0
httpx==0.27.0
passlib[bcrypt]==1.7.4
reportlab==4.1.0
python-multipart==0.0.9
bcrypt==4.0.1
